* release/40.x:
  #6032 - Scrolling to annotation no longer works

+++ inception/inception-brat-editor/src/main/ts/build.mjs
@@ -62,7 +62,9 @@ const defaults = {
 }
 
 fs.mkdirsSync(`${outbase}`)
-fs.emptyDirSync(outbase)
+if (!argv.live) {
+  fs.emptyDirSync(outbase)
+}
 
 if (argv.live) {
   const context1 = await esbuild.context(Object.assign({
