+++ inception/inception-html-apache-annotator-editor/src/main/ts/build.mjs
@@ -67,7 +67,9 @@ const defaults = {
 }
 
 fs.mkdirsSync(`${outbase}`)
-fs.emptyDirSync(outbase)
+if (!argv.live) {
+  fs.emptyDirSync(outbase)
+}
 
 if (argv.live) {
   const context = await esbuild.context({
