+++ inception/inception-ui-scheduling/src/main/ts/build.mjs
@@ -51,7 +51,9 @@ const defaults = {
 }
 
 fs.mkdirsSync(`${outbase}`)
-fs.emptyDirSync(outbase)
+if (!argv.live) {
+  fs.emptyDirSync(outbase)
+}
 
 if (argv.live) {
   const context = await esbuild.context(defaults)
